Parse all time points and spans as float64
refs #112

+++ configobject/objecttypes/checkcommand/checkcommand.go
@@ -32,7 +32,7 @@ type CheckCommand struct {
 	NameCi             *string `json:"name_ci"`
 	ZoneId             string  `json:"zone_id"`
 	Command            string  `json:"command"`
-	Timeout            float32 `json:"timeout"`
+	Timeout            float64 `json:"timeout"`
 }
 
 func NewCheckCommand() connection.Row {
